pyproject.toml
[tool.poetry]
name = "email-targeting"
version = "0.1.0"
description = ""
authors = ["Noah Jaffe <noah.h.jaffe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.2"
langchain = "^0.2.1"
langchain-openai = "^0.1.7"
langchain-community = "^0.2.1"
openpyxl = "^3.1.2"
faiss-cpu = "^1.8.0"
bertopic = "^0.8.1"
torch = { version = "=1.9.0+cu111", source = "pytorch" }
torchvision = { version = "=0.10.0+cu111", source = "pytorch" }

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/torch_stable.html"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]